Move PR event handlers to a separate module
