+++ src/bors/handlers/pr_events.rs
@@ -1,0 +1,217 @@
+use crate::bors::event::{PullRequestEdited, PullRequestPushed};
+use crate::bors::handlers::labels::handle_label_trigger;
+use crate::bors::{Comment, RepositoryState};
+use crate::github::{CommitSha, LabelTrigger, PullRequestNumber};
+use crate::PgDbClient;
+use std::sync::Arc;
+
+pub(super) async fn handle_pull_request_edited(
+    repo_state: Arc<RepositoryState>,
+    db: Arc<PgDbClient>,
+    payload: PullRequestEdited,
+) -> anyhow::Result<()> {
+    // If the base branch has changed, unapprove the PR
+    let Some(_) = payload.from_base_sha else {
+        return Ok(());
+    };
+
+    let pr_model = db
+        .get_or_create_pull_request(repo_state.repository(), payload.pull_request.number)
+        .await?;
+    if !pr_model.is_approved() {
+        return Ok(());
+    }
+
+    let pr_number = payload.pull_request.number;
+    db.unapprove(repo_state.repository(), pr_number).await?;
+    handle_label_trigger(&repo_state, pr_number, LabelTrigger::Unapproved).await?;
+    notify_of_edited_pr(&repo_state, pr_number, &payload.pull_request.base.name).await
+}
+
+pub(super) async fn handle_push_to_pull_request(
+    repo_state: Arc<RepositoryState>,
+    db: Arc<PgDbClient>,
+    payload: PullRequestPushed,
+) -> anyhow::Result<()> {
+    let pr = &payload.pull_request;
+    let pr_model = db
+        .get_or_create_pull_request(repo_state.repository(), pr.number)
+        .await?;
+
+    if !pr_model.is_approved() {
+        return Ok(());
+    }
+
+    let pr_number = pr_model.number;
+    db.unapprove(repo_state.repository(), pr_number).await?;
+    handle_label_trigger(&repo_state, pr_number, LabelTrigger::Unapproved).await?;
+    notify_of_pushed_pr(&repo_state, pr_number, pr.head.sha.clone()).await
+}
+
+async fn notify_of_edited_pr(
+    repo: &RepositoryState,
+    pr_number: PullRequestNumber,
+    base_name: &str,
+) -> anyhow::Result<()> {
+    repo.client
+        .post_comment(
+            pr_number,
+            Comment::new(format!(
+                r#":warning: The base branch changed to `{base_name}`, and the
+PR will need to be re-approved."#,
+            )),
+        )
+        .await
+}
+
+async fn notify_of_pushed_pr(
+    repo: &RepositoryState,
+    pr_number: PullRequestNumber,
+    head_sha: CommitSha,
+) -> anyhow::Result<()> {
+    repo.client
+        .post_comment(
+            pr_number,
+            Comment::new(format!(
+                r#":warning: A new commit `{}` was pushed to the branch, the
+PR will need to be re-approved."#,
+                head_sha
+            )),
+        )
+        .await
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::tests::mocks::{
+        assert_pr_approved_by, assert_pr_unapproved, create_world_with_approve_config,
+        default_pr_number, run_test, BorsBuilder, PullRequestChangeEvent, User,
+    };
+
+    #[sqlx::test]
+    async fn unapprove_on_base_edited(pool: sqlx::PgPool) {
+        BorsBuilder::new(pool)
+            .world(create_world_with_approve_config())
+            .run_test(|mut tester| async {
+                tester.post_comment("@bors r+").await?;
+                assert_eq!(
+                    tester.get_comment().await?,
+                    format!(
+                        "Commit pr-{}-sha has been approved by `{}`",
+                        default_pr_number(),
+                        User::default_user().name
+                    ),
+                );
+                tester
+                    .edit_pull_request(
+                        default_pr_number(),
+                        PullRequestChangeEvent {
+                            from_base_sha: Some("main-sha".to_string()),
+                        },
+                    )
+                    .await?;
+
+                assert_eq!(
+                    tester.get_comment().await?,
+                    r#":warning: The base branch changed to `main`, and the
+PR will need to be re-approved."#,
+                );
+                assert_pr_unapproved(&tester, default_pr_number().into()).await;
+                Ok(tester)
+            })
+            .await;
+    }
+
+    #[sqlx::test]
+    async fn edit_pr_do_nothing_when_base_not_edited(pool: sqlx::PgPool) {
+        BorsBuilder::new(pool)
+            .world(create_world_with_approve_config())
+            .run_test(|mut tester| async {
+                tester.post_comment("@bors r+").await?;
+                assert_eq!(
+                    tester.get_comment().await?,
+                    format!(
+                        "Commit pr-{}-sha has been approved by `{}`",
+                        default_pr_number(),
+                        User::default_user().name
+                    ),
+                );
+                tester
+                    .edit_pull_request(
+                        default_pr_number(),
+                        PullRequestChangeEvent {
+                            from_base_sha: None,
+                        },
+                    )
+                    .await?;
+
+                assert_pr_approved_by(
+                    &tester,
+                    default_pr_number().into(),
+                    &User::default_user().name,
+                )
+                .await;
+                Ok(tester)
+            })
+            .await;
+    }
+
+    #[sqlx::test]
+    async fn edit_pr_do_nothing_when_not_approved(pool: sqlx::PgPool) {
+        run_test(pool, |mut tester| async {
+            tester
+                .edit_pull_request(
+                    default_pr_number(),
+                    PullRequestChangeEvent {
+                        from_base_sha: Some("main-sha".to_string()),
+                    },
+                )
+                .await?;
+
+            // No comment should be posted
+            Ok(tester)
+        })
+        .await;
+    }
+
+    #[sqlx::test]
+    async fn unapprove_on_push(pool: sqlx::PgPool) {
+        BorsBuilder::new(pool)
+            .world(create_world_with_approve_config())
+            .run_test(|mut tester| async {
+                tester.post_comment("@bors r+").await?;
+                assert_eq!(
+                    tester.get_comment().await?,
+                    format!(
+                        "Commit pr-{}-sha has been approved by `{}`",
+                        default_pr_number(),
+                        User::default_user().name
+                    ),
+                );
+                tester.push_to_pull_request(default_pr_number()).await?;
+
+                assert_eq!(
+                    tester.get_comment().await?,
+                    format!(
+                        r#":warning: A new commit `pr-{}-sha` was pushed to the branch, the
+PR will need to be re-approved."#,
+                        default_pr_number()
+                    )
+                );
+                assert_pr_unapproved(&tester, default_pr_number().into()).await;
+                Ok(tester)
+            })
+            .await;
+    }
+
+    #[sqlx::test]
+    async fn push_to_pr_do_nothing_when_not_approved(pool: sqlx::PgPool) {
+        run_test(pool, |mut tester| async {
+            tester.push_to_pull_request(default_pr_number()).await?;
+
+            // No comment should be posted
+            Ok(tester)
+        })
+        .await;
+    }
+}
