+++ src/bors/handlers/review.rs
@@ -2,18 +2,14 @@ use std::sync::Arc;
 
 use crate::bors::command::Approver;
 use crate::bors::command::RollupMode;
-use crate::bors::event::PullRequestEdited;
-use crate::bors::event::PullRequestPushed;
 use crate::bors::handlers::deny_request;
 use crate::bors::handlers::has_permission;
 use crate::bors::handlers::labels::handle_label_trigger;
 use crate::bors::Comment;
 use crate::bors::RepositoryState;
-use crate::github::CommitSha;
 use crate::github::GithubUser;
 use crate::github::LabelTrigger;
 use crate::github::PullRequest;
-use crate::github::PullRequestNumber;
 use crate::permissions::PermissionType;
 use crate::PgDbClient;
 
@@ -134,53 +130,19 @@ pub(super) async fn command_set_rollup(
         .await
 }
 
-pub(super) async fn handle_pull_request_edited(
-    repo_state: Arc<RepositoryState>,
-    db: Arc<PgDbClient>,
-    payload: PullRequestEdited,
-) -> anyhow::Result<()> {
-    // If the base branch has changed, unapprove the PR
-    let Some(_) = payload.from_base_sha else {
-        return Ok(());
-    };
-
-    let pr_model = db
-        .get_or_create_pull_request(repo_state.repository(), payload.pull_request.number)
-        .await?;
-    if !pr_model.is_approved() {
-        return Ok(());
-    }
-
-    let pr_number = payload.pull_request.number;
-    db.unapprove(repo_state.repository(), pr_number).await?;
-    handle_label_trigger(&repo_state, pr_number, LabelTrigger::Unapproved).await?;
-    notify_of_edited_pr(&repo_state, pr_number, &payload.pull_request.base.name).await
-}
-
-pub(super) async fn handle_push_to_pull_request(
-    repo_state: Arc<RepositoryState>,
-    db: Arc<PgDbClient>,
-    payload: PullRequestPushed,
-) -> anyhow::Result<()> {
-    let pr = &payload.pull_request;
-    let pr_model = db
-        .get_or_create_pull_request(repo_state.repository(), pr.number)
-        .await?;
-
-    if !pr_model.is_approved() {
-        return Ok(());
-    }
-
-    let pr_number = pr_model.number;
-    db.unapprove(repo_state.repository(), pr_number).await?;
-    handle_label_trigger(&repo_state, pr_number, LabelTrigger::Unapproved).await?;
-    notify_of_pushed_pr(&repo_state, pr_number, pr.head.sha.clone()).await
-}
-
 fn sufficient_delegate_permission(repo: Arc<RepositoryState>, author: &GithubUser) -> bool {
     repo.permissions
         .load()
         .has_permission(author.id, PermissionType::Review)
+}
+
+async fn notify_of_unapproval(repo: &RepositoryState, pr: &PullRequest) -> anyhow::Result<()> {
+    repo.client
+        .post_comment(
+            pr.number,
+            Comment::new(format!("Commit {} has been unapproved", pr.head.sha)),
+        )
+        .await
 }
 
 async fn notify_of_approval(
@@ -194,48 +156,6 @@ async fn notify_of_approval(
             Comment::new(format!(
                 "Commit {} has been approved by `{}`",
                 pr.head.sha, approver
-            )),
-        )
-        .await
-}
-
-async fn notify_of_unapproval(repo: &RepositoryState, pr: &PullRequest) -> anyhow::Result<()> {
-    repo.client
-        .post_comment(
-            pr.number,
-            Comment::new(format!("Commit {} has been unapproved", pr.head.sha)),
-        )
-        .await
-}
-
-async fn notify_of_edited_pr(
-    repo: &RepositoryState,
-    pr_number: PullRequestNumber,
-    base_name: &str,
-) -> anyhow::Result<()> {
-    repo.client
-        .post_comment(
-            pr_number,
-            Comment::new(format!(
-                r#":warning: The base branch changed to `{base_name}`, and the
-PR will need to be re-approved."#,
-            )),
-        )
-        .await
-}
-
-async fn notify_of_pushed_pr(
-    repo: &RepositoryState,
-    pr_number: PullRequestNumber,
-    head_sha: CommitSha,
-) -> anyhow::Result<()> {
-    repo.client
-        .post_comment(
-            pr_number,
-            Comment::new(format!(
-                r#":warning: A new commit `{}` was pushed to the branch, the
-PR will need to be re-approved."#,
-                head_sha
             )),
         )
         .await
@@ -256,16 +176,18 @@ async fn notify_of_delegation(
 
 #[cfg(test)]
 mod tests {
+    use crate::tests::mocks::{
+        assert_pr_approved_by, assert_pr_unapproved, create_world_with_approve_config,
+    };
     use crate::{
         bors::{
             handlers::{trybuild::TRY_MERGE_BRANCH_NAME, TRY_BRANCH_NAME},
             RollupMode,
         },
-        github::PullRequestNumber,
         permissions::PermissionType,
         tests::mocks::{
-            default_pr_number, default_repo_name, run_test, BorsBuilder, BorsTester, Comment,
-            Permissions, PullRequestChangeEvent, User, World,
+            default_pr_number, default_repo_name, run_test, BorsBuilder, Comment, Permissions,
+            User, World,
         },
     };
 
@@ -291,7 +213,7 @@ mod tests {
                     })
                     .await?;
 
-                check_pr_approved_by(
+                assert_pr_approved_by(
                     &tester,
                     default_pr_number().into(),
                     &User::default_user().name,
@@ -319,7 +241,7 @@ mod tests {
                     ),
                 );
 
-                check_pr_approved_by(&tester, default_pr_number().into(), approve_user).await;
+                assert_pr_approved_by(&tester, default_pr_number().into(), approve_user).await;
                 Ok(tester)
             })
             .await;
@@ -370,7 +292,7 @@ mod tests {
                         User::default_user().name
                     ),
                 );
-                check_pr_approved_by(
+                assert_pr_approved_by(
                     &tester,
                     default_pr_number().into(),
                     &User::default_user().name,
@@ -381,176 +303,10 @@ mod tests {
                     tester.get_comment().await?,
                     format!("Commit pr-{}-sha has been unapproved", default_pr_number()),
                 );
-                check_pr_unapproved(&tester, default_pr_number().into()).await;
+                assert_pr_unapproved(&tester, default_pr_number().into()).await;
                 Ok(tester)
             })
             .await;
-    }
-
-    #[sqlx::test]
-    async fn unapprove_on_base_edited(pool: sqlx::PgPool) {
-        BorsBuilder::new(pool)
-            .world(create_world_with_approve_config())
-            .run_test(|mut tester| async {
-                tester.post_comment("@bors r+").await?;
-                assert_eq!(
-                    tester.get_comment().await?,
-                    format!(
-                        "Commit pr-{}-sha has been approved by `{}`",
-                        default_pr_number(),
-                        User::default_user().name
-                    ),
-                );
-                tester
-                    .edit_pull_request(
-                        default_pr_number(),
-                        PullRequestChangeEvent {
-                            from_base_sha: Some("main-sha".to_string()),
-                        },
-                    )
-                    .await?;
-
-                assert_eq!(
-                    tester.get_comment().await?,
-                    r#":warning: The base branch changed to `main`, and the
-PR will need to be re-approved."#,
-                );
-                check_pr_unapproved(&tester, default_pr_number().into()).await;
-                Ok(tester)
-            })
-            .await;
-    }
-
-    #[sqlx::test]
-    async fn edit_pr_do_nothing_when_base_not_edited(pool: sqlx::PgPool) {
-        BorsBuilder::new(pool)
-            .world(create_world_with_approve_config())
-            .run_test(|mut tester| async {
-                tester.post_comment("@bors r+").await?;
-                assert_eq!(
-                    tester.get_comment().await?,
-                    format!(
-                        "Commit pr-{}-sha has been approved by `{}`",
-                        default_pr_number(),
-                        User::default_user().name
-                    ),
-                );
-                tester
-                    .edit_pull_request(
-                        default_pr_number(),
-                        PullRequestChangeEvent {
-                            from_base_sha: None,
-                        },
-                    )
-                    .await?;
-
-                check_pr_approved_by(
-                    &tester,
-                    default_pr_number().into(),
-                    &User::default_user().name,
-                )
-                .await;
-                Ok(tester)
-            })
-            .await;
-    }
-
-    #[sqlx::test]
-    async fn edit_pr_do_nothing_when_not_approved(pool: sqlx::PgPool) {
-        run_test(pool, |mut tester| async {
-            tester
-                .edit_pull_request(
-                    default_pr_number(),
-                    PullRequestChangeEvent {
-                        from_base_sha: Some("main-sha".to_string()),
-                    },
-                )
-                .await?;
-
-            // No comment should be posted
-            Ok(tester)
-        })
-        .await;
-    }
-
-    #[sqlx::test]
-    async fn unapprove_on_push(pool: sqlx::PgPool) {
-        BorsBuilder::new(pool)
-            .world(create_world_with_approve_config())
-            .run_test(|mut tester| async {
-                tester.post_comment("@bors r+").await?;
-                assert_eq!(
-                    tester.get_comment().await?,
-                    format!(
-                        "Commit pr-{}-sha has been approved by `{}`",
-                        default_pr_number(),
-                        User::default_user().name
-                    ),
-                );
-                tester.push_to_pull_request(default_pr_number()).await?;
-
-                assert_eq!(
-                    tester.get_comment().await?,
-                    format!(
-                        r#":warning: A new commit `pr-{}-sha` was pushed to the branch, the
-PR will need to be re-approved."#,
-                        default_pr_number()
-                    )
-                );
-                check_pr_unapproved(&tester, default_pr_number().into()).await;
-                Ok(tester)
-            })
-            .await;
-    }
-
-    #[sqlx::test]
-    async fn push_to_pr_do_nothing_when_not_approved(pool: sqlx::PgPool) {
-        run_test(pool, |mut tester| async {
-            tester.push_to_pull_request(default_pr_number()).await?;
-
-            // No comment should be posted
-            Ok(tester)
-        })
-        .await;
-    }
-
-    fn create_world_with_approve_config() -> World {
-        let world = World::default();
-        world.default_repo().lock().set_config(
-            r#"
-[labels]
-approve = ["+approved"]
-"#,
-        );
-        world
-    }
-
-    async fn check_pr_approved_by(
-        tester: &BorsTester,
-        pr_number: PullRequestNumber,
-        approved_by: &str,
-    ) {
-        let pr_in_db = tester
-            .db()
-            .get_or_create_pull_request(&default_repo_name(), pr_number)
-            .await
-            .unwrap();
-        assert_eq!(pr_in_db.approved_by, Some(approved_by.to_string()));
-        let repo = tester.default_repo();
-        let pr = repo.lock().get_pr(default_pr_number()).clone();
-        pr.check_added_labels(&["approved"]);
-    }
-
-    async fn check_pr_unapproved(tester: &BorsTester, pr_number: PullRequestNumber) {
-        let pr_in_db = tester
-            .db()
-            .get_or_create_pull_request(&default_repo_name(), pr_number)
-            .await
-            .unwrap();
-        assert_eq!(pr_in_db.approved_by, None);
-        let repo = tester.default_repo();
-        let pr = repo.lock().get_pr(default_pr_number()).clone();
-        pr.check_removed_labels(&["approved"]);
     }
 
     #[sqlx::test]
@@ -699,7 +455,7 @@ approve = ["+approved"]
                 tester.post_comment("@bors r+").await?;
                 tester.expect_comments(1).await;
 
-                check_pr_approved_by(&tester, default_pr_number().into(), &User::default().name)
+                assert_pr_approved_by(&tester, default_pr_number().into(), &User::default().name)
                     .await;
                 Ok(tester)
             })
@@ -853,7 +609,7 @@ approve = ["+approved"]
                     .post_comment(Comment::from("@bors r+").with_author(User::default()))
                     .await?;
                 tester.expect_comments(1).await;
-                check_pr_approved_by(&tester, default_pr_number().into(), &User::default().name)
+                assert_pr_approved_by(&tester, default_pr_number().into(), &User::default().name)
                     .await;
 
                 tester.post_comment(as_reviewer("@bors r-")).await?;
